greenhouse_system/requirements.txt
asyncua==1.1.8
mysql-connector-python
dash
plotly
